+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- [Result] update format from map `%{communities: [community], bridges: [birdge]}` to tuple `{[community], [bridge]}`.
+
 ## 0.2.0 - 2025-09-12
 
 ### Added